+++ CHANGELOG.md
@@ -105,6 +105,10 @@ There are following change types:
 
 - Ensure preview window is never hidden, even if cursor is on the line for a not (yet) existing file system entry. This reduces flickering of preview window when creating new files in Insert mode.
 
+### Expand
+
+- Add `config.content.highlight` to customize how file system entry is highlighted. Defaults to a new `default_highlight()` function.
+
 ## mini.hues
 
 ### Evolve